server/pyproject.toml
[tool.poetry]
name = "itfox"
version = "0.1.0"
description = ""
authors = ["Roman Pavlyuchenko <romanpavlyuchenko@inbox.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
flake8 = "^6.0.0"
djangorestframework-simplejwt = "^5.2.2"
django-debug-toolbar = "^3.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
